chore(submitter): remove dumpPrivKey (#15)
This pull request removes the usage of the dumpPrivKey RPC method from
the submitter and end-to-end (E2E) tests. The dumpPrivKey method is
specific to legacy wallets and is no longer supported in the latest
versions of bitcoind. Eliminating its usage is a necessary step for
upgrading to newer versions of bitcoind, as its deprecation would
otherwise hinder the upgrade process.

Updates the version of bitcoind to v27.0

References
[issue](https://github.com/babylonchain/vigilante/issues/227).